+++ climt/__init__.py
@@ -7,12 +7,12 @@ from ._core import (
     get_interface_values,
     climt_quantity_descriptions,
     bolton_q_sat, bolton_dqsat_dT,
-    ClimtImplicitPrognostic, ClimtPrognostic, ClimtImplicit)
+    ClimtImplicitPrognostic, ClimtPrognostic, ClimtImplicit, ClimtDiagnostic)
 
 from ._components import (
     Frierson06LongwaveOpticalDepth, GrayLongwaveRadiation, HeldSuarez,
     GridScaleCondensation, BergerSolarInsolation, SimplePhysics, RRTMGLongwave, RRTMGShortwave,
-    EmanuelConvection, SlabSurface)
+    EmanuelConvection, SlabSurface, IceSheet)
 
 sympl.add_direction_names(z=['interface_levels', 'mid_levels'])
 
@@ -20,9 +20,9 @@ __all__ = (
     get_default_state, mass_to_volume_mixing_ratio,
     get_interface_values,
     bolton_q_sat, bolton_dqsat_dT, climt_quantity_descriptions,
-    ClimtImplicitPrognostic, ClimtPrognostic, ClimtImplicit,
+    ClimtImplicitPrognostic, ClimtPrognostic, ClimtImplicit, ClimtDiagnostic,
     Frierson06LongwaveOpticalDepth, GrayLongwaveRadiation, HeldSuarez,
     GridScaleCondensation, BergerSolarInsolation, SimplePhysics, RRTMGLongwave, RRTMGShortwave,
-    EmanuelConvection, SlabSurface)
+    EmanuelConvection, SlabSurface, IceSheet)
 
 __version__ = '0.9'