+++ climt/_components/__init__.py
@@ -6,8 +6,9 @@ from .simple_physics import SimplePhysics
 from .rrtmg import RRTMGLongwave, RRTMGShortwave
 from .emanuel import EmanuelConvection
 from .slab_surface import SlabSurface
+from .surface_ice import IceSheet
 
 __all__ = (
     Frierson06LongwaveOpticalDepth, GrayLongwaveRadiation,
     HeldSuarez, GridScaleCondensation, BergerSolarInsolation, SimplePhysics,
-    RRTMGLongwave, RRTMGShortwave, EmanuelConvection, SlabSurface)
+    RRTMGLongwave, RRTMGShortwave, EmanuelConvection, SlabSurface, IceSheet)