+++ HISTORY.rst
@@ -4,6 +4,8 @@ History
 
 Latest
 ------
+* Arrays can now have arbitrary dtype (to use qualitative, string, quantities)
+* Added Emanuel Convection, surface energy balance model and ice sheet energy balance
 * 2D coordinates are now supported for horizontal coordinates
 * Replaced create_output_arrays() with a more general
   get_state_dict_for() and get_numpy_arrays_from_state()